New class to accommodate savings/investing account
New class to accommodate savings/investing account and features. See comments in code highlighting the approach to using a class. Better approaches may exist but this can serve as a starting point. More to be added to this class and to NewBank.java to incorporate it into src

This cater to FR10 and Fr19 among others that target savings/investing features

The basic functionality is to eventually have the option to deposit money into the investing account and earn interest on that money based on the type of investment account selected (three options are suggested in the comments) 

Next up:
- add util.calendar or similar to grow start amount by the rate of return each year
- create doubles for each investment account type and its related rate of return
- let client choose account type and start earning interest